Improve tests
- Add testing for run() function (function was moved to its own run.js
  file)
- Split unit tests in different files
