+++ approval.test.js
@@ -1,114 +1,5 @@
-const inputs = require('./inputs');
-const review = require('./review');
 const approval = require('./approval');
 const owners = require('./owners');
-
-describe ('Parse Owners', () => {
-    test('all keys', async () => {
-        const {maintainers, areaReviewers, areaApprovers} = inputs.parseOwners('testdata/owners.yml');
-        expect(maintainers).toEqual(['ted']);
-        expect(areaReviewers.get('area/area-1')).toEqual(expect.arrayContaining(['alice', 'bob']));
-        expect(areaReviewers.get('area/area-2')).toEqual(expect.arrayContaining(['bob', 'john']));
-        expect(areaApprovers.get('area/area-1')).toEqual(expect.arrayContaining(['alice', 'bob']));
-        expect(areaApprovers.get('area/area-2')).toEqual(expect.arrayContaining(['bob', 'john']));
-    });
-
-    test('missing approvers', async () => {
-        const {maintainers, areaReviewers, areaApprovers} = inputs.parseOwners('testdata/owners-without-approvers.yml');
-        expect(maintainers).toEqual(['ted']);
-        expect(areaReviewers.get('area/area-1')).toEqual(expect.arrayContaining(['alice', 'bob']));
-        expect(areaReviewers.get('area/area-2')).toEqual(expect.arrayContaining(['bob', 'john']));
-        expect(areaApprovers).toEqual(new Map());
-    });
-
-    test('missing maintainers', async () => {
-        const {maintainers, areaReviewers, areaApprovers} = inputs.parseOwners('testdata/owners-without-maintainers.yml');
-        expect(maintainers).toEqual([]);
-        expect(areaReviewers.get('area/area-1')).toEqual(expect.arrayContaining(['alice', 'bob']));
-        expect(areaReviewers.get('area/area-2')).toEqual(expect.arrayContaining(['bob', 'john']));
-        expect(areaApprovers.get('area/area-1')).toEqual(expect.arrayContaining(['alice', 'bob']));
-        expect(areaApprovers.get('area/area-2')).toEqual(expect.arrayContaining(['bob', 'john']));
-    });
-});
-
-describe('Compute reviewers', () => {
-    let labels;
-    let author;
-    let maintainers;
-    let reviewers;
-    let approvers;
-    let requestReviewsFromMaintainersIfNeeded;
-
-    function computeReviewers() {
-        const areaReviewers = new Map(reviewers);
-        const areaApprovers = new Map(approvers);
-        const config = {
-            minApprovingReviewsPerArea: 1,
-            maintainers: maintainers,
-            areaReviewers: areaReviewers,
-            areaApprovers: areaApprovers,
-            areaReviewersRegexList: owners.buildRegexList(areaReviewers),
-            areaApproversRegexList: owners.buildRegexList(areaApprovers),
-            requestReviewsFromMaintainersIfNeeded: requestReviewsFromMaintainersIfNeeded,
-        };
-        return review.computeReviewers(labels, author, config);
-    }
-
-    beforeEach(() => {
-        labels = ['documentation'];
-        author = "alice";
-        maintainers = ['alice'];
-        reviewers = [];
-        approvers = [];
-        requestReviewsFromMaintainersIfNeeded = false;
-    });
-
-    test('author removal', async () => {
-        reviewers = [['documentation', ['alice', 'bob']]];
-        expect(computeReviewers()).toEqual(['bob']);
-    });
-
-    test('not a known label', async() => {
-        reviewers = [['not-documentation', ['alice', 'bob']]];
-        expect(computeReviewers()).toEqual([]);
-    });
-
-    test('combine 1', async() => {
-        author = "clara";
-        reviewers = [['documentation', ['alice', 'bob']]];
-        approvers = [['documentation', ['alice']]];
-        expect(computeReviewers()).toEqual(expect.arrayContaining(['alice', 'bob']));
-    });
-
-    test('combine 2', async() => {
-        author = "clara";
-        reviewers = [['documentation', ['alice']]];
-        approvers = [['documentation', ['alice', 'bob']]];
-        expect(computeReviewers()).toEqual(expect.arrayContaining(['alice', 'bob']));
-    });
-
-    test('regex', async() => {
-        reviewers = [['doc*', ['alice', 'bob']]];
-        expect(computeReviewers()).toEqual(['bob']);
-    });
-
-    test('regex precedence', async() => {
-        reviewers = [['documentation', ['mike']], ['doc*', ['alice', 'bob']]];
-        expect(computeReviewers()).toEqual(['mike']);
-    });
-
-    test('request reviews from maintainers if needed - default', async() => {
-        reviewers = [['documentation', ['bob']]];
-        expect(computeReviewers()).toEqual(['bob']);
-    });
-
-    test('request reviews from maintainers if needed - true', async() => {
-        requestReviewsFromMaintainersIfNeeded = true;
-        maintainers = ['mike'];
-        reviewers = [['documentation', ['bob']]];
-        expect(computeReviewers()).toEqual(expect.arrayContaining(['bob', 'mike']));
-    });
-});
 
 describe('PR can be merged', () => {
     let labels;
